wiobdos() の戻り値の型を uint16 に変更

+++ Wio_RunCPM_vt100_UsbKB/hardware/arduino/wioterm.h
@@ -15,7 +15,7 @@ QueueHandle_t xQueue;
 RTC_SAMD51 rtc;
 DateTime now;
 
-uint8 wiobdos(uint16 dmaaddr) {
+uint16 wiobdos(uint16 dmaaddr) {
 	return(0x00);
 }
